dom/serviceworkers/test/test_register_https_in_http.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 1172948 - Test that registering a service worker from inside an HTTPS iframe embedded in an HTTP iframe doesn't work</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">

  function runTest() {
    var iframe = document.createElement("iframe");
    iframe.src = "https://example.com/tests/dom/serviceworkers/test/register_https.html";
    document.body.appendChild(iframe);

    window.onmessage = event => {
      switch (event.data.type) {
      case "ok":
        ok(event.data.status, event.data.msg);
        break;
      case "done":
        SimpleTest.finish();
        break;
      }
    };
  }

  SimpleTest.waitForExplicitFinish();
  onload = function() {
    SpecialPowers.pushPrefEnv({"set": [
      ["dom.serviceWorkers.exemptFromPerDomainMax", true],
      ["dom.serviceWorkers.enabled", true],
    ]}, runTest);
  };
</script>
</body>
</html>
